rename imageresize to resizeimage
