+++ resizeimage/imageexceptions.py
