+++ tox.ini
@@ -10,8 +10,8 @@ setenv=
 
 [testenv:py27]
 commands=
-	coverage run --source imageresize setup.py test
+	coverage run --source resizeimage setup.py test
 
 [testenv:py34]
 commands=
-	coverage run --source imageresize setup.py test
+	coverage run --source resizeimage setup.py test
